+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * refactor: use `Vec::retain` for more idiomatic and efficient collection cleanup in `src/run.rs`.
+* refactor: improve readability of regex capture group selection logic in `src/run.rs`.
 
 
 ## [0.2.1] (2026-05-19)